Make move player-state set to internal
Was previously done over the instance but required player to be pub
This moves the player into its own arc and thus allows this to be set
on events directly.
This doesn't fully solve the problem of having to leave some knowledge
about the idea of instances inside the event chain but allows to
modularize it more.

Signed-off-by: Aron Heinecke <aron.heinecke@t-online.de>